+++ dev/src/crossroads.js
@@ -5,7 +5,7 @@
     /**
      * @constructor
      */
-    function Crossroads(){
+    function Crossroads() {
         this._routes = [];
         this.bypassed = new signals.Signal();
         this.routed = new signals.Signal();
@@ -13,49 +13,63 @@
 
     Crossroads.prototype = {
 
-        create : function(){
+        normalizeFn : null,
+
+        create : function () {
             return new Crossroads();
         },
 
         shouldTypecast : false,
 
-        addRoute : function(pattern, callback, priority){
+        addRoute : function (pattern, callback, priority) {
             var route = new Route(pattern, callback, priority, this);
             this._sortedInsert(route);
             return route;
         },
 
-        removeRoute : function(route){
+        removeRoute : function (route) {
             var i = arrayIndexOf(this._routes, route);
-            if(i >= 0) this._routes.splice(i, 1);
+            if (i !== -1) {
+                this._routes.splice(i, 1);
+            }
             route._destroy();
         },
 
-        removeAllRoutes : function(){
+        removeAllRoutes : function () {
             var n = this.getNumRoutes();
-            while(n--){
+            while (n--) {
                 this._routes[n]._destroy();
             }
             this._routes.length = 0;
         },
 
-        parse : function(request){
+        parse : function (request) {
             request = request || '';
-            var route = this._getMatchedRoute(request),
-                params = route? route._getParamsArray(request) : null;
-            if(route){
-                params? route.matched.dispatch.apply(route.matched, params) : route.matched.dispatch();
-                this.routed.dispatch(request, route, params);
-            }else{
+
+            var routes = this._getMatchedRoutes(request),
+                i = 0,
+                n = routes.length,
+                cur;
+
+            if (n) {
+                //shold be incremental loop, execute routes in order
+                while (i < n) {
+                    cur = routes[i];
+                    cur.route.matched.dispatch.apply(cur.route.matched, cur.params);
+                    cur.isFirst = !i;
+                    this.routed.dispatch(request, cur);
+                    i += 1;
+                }
+            } else {
                 this.bypassed.dispatch(request);
             }
         },
 
-        getNumRoutes : function(){
+        getNumRoutes : function () {
             return this._routes.length;
         },
 
-        _sortedInsert : function(route){
+        _sortedInsert : function (route) {
             //simplified insertion sort
             var routes = this._routes,
                 n = routes.length;
@@ -63,17 +77,24 @@
             routes.splice(n+1, 0, route);
         },
 
-        _getMatchedRoute : function(request){
-            var routes = this._routes,
+        _getMatchedRoutes : function (request) {
+            var res = [],
+                routes = this._routes,
                 n = routes.length,
                 route;
-            while(route = routes[--n]){ //should be decrement loop since higher priorities are added at the end of array  
-                if(route.match(request)) return route;
+            //should be decrement loop since higher priorities are added at the end of array
+            while (route = routes[--n]) {
+                if ((!res.length || route.greedy) && route.match(request)) {
+                    res.push({
+                        route : route,
+                        params : route._getParamsArray(request)
+                    });
+                }
             }
-            return null;
+            return res;
         },
 
-        toString : function(){
+        toString : function () {
             return '[crossroads numRoutes:'+ this.getNumRoutes() +']';
         }
     };