+++ dev/tests/spec/signals.spec.js
@@ -10,6 +10,8 @@ describe('crossroads Signals', function(){
 
     afterEach(function(){
         crossroads.removeAllRoutes();
+        crossroads.bypassed.removeAll();
+        crossroads.routed.removeAll();
     });
 
 
@@ -34,31 +36,34 @@ describe('crossroads Signals', function(){
         expect( count ).toBe( 2 );
     });
 
+
     it('should dispatch routed at each match', function(){
-        var count = 0, 
-            requests = [], 
+        var count = 0,
+            requests = [],
             count2 = 0,
-            routed;
+            routed,
+            first;
 
         var a = crossroads.addRoute('/{foo}_{bar}');
         a.matched.add(function(foo, bar){
-            count2++;
+            count++;
         });
 
         crossroads.bypassed.add(function(request){
             requests.push(request);
-            count++;
+            count2++;
         });
 
-        crossroads.routed.add(function(request, route, params){
+        crossroads.routed.add(function(request, data){
             requests.push(request);
             count++;
 
             expect( request ).toBe( '/foo_bar' );
-            expect( route ).toBe( a );
-            expect( params[0] ).toEqual( 'foo' );
-            expect( params[1] ).toEqual( 'bar' );
+            expect( data.route ).toBe( a );
+            expect( data.params[0] ).toEqual( 'foo' );
+            expect( data.params[1] ).toEqual( 'bar' );
             routed = true;
+            first = data.isFirst;
         });
 
         crossroads.parse('/lorem/ipsum');
@@ -69,10 +74,33 @@ describe('crossroads Signals', function(){
         expect( count ).toBe( 2 );
         expect( count2 ).toBe( 1 );
         expect( routed ).toEqual( true );
+        expect( first ).toEqual( true );
+
+    });
 
 
-        crossroads.bypassed.removeAll();
-        crossroads.routed.removeAll();
+    it('isFirst should be false on greedy matches', function () {
+
+        var count = 0,
+            firsts = [];
+
+        crossroads.routed.add(function(req, data){
+            count += 1;
+            firsts.push(data.isFirst);
+        });
+
+        //anti-pattern!
+        crossroads.addRoute('/{a}/{b}');
+        crossroads.addRoute('/{a}/{b}').greedy = true;
+        crossroads.addRoute('/{a}/{b}').greedy = true;
+
+        crossroads.parse('/foo/bar');
+
+        expect( count ).toEqual( 3 );
+        expect( firsts[0] ).toEqual( true );
+        expect( firsts[1] ).toEqual( false );
+        expect( firsts[2] ).toEqual( false );
+
     });
 
 });